remove pub from context vars, remove Bot, rename bot.rs to steps.rs, implement Worker
